Added GreenSurvivors as authors dokumentation renamed blockRecorder package to just recorder to reflect that entities get recorded as well. Also, it makes codacy happy.
