--- Minigames/src/main/java/au/com/mineauz/minigames/blockRecorder/package-info.java
@@ -1,4 +1,0 @@
-/**
- * BlockRecorder is the system for recording block / entity changes in a game and rolling them back afterward.
- */
-package au.com.mineauz.minigames.blockRecorder;
