+++ Minigames/src/main/java/au/com/mineauz/minigames/commands/ICommand.java
@@ -7,16 +7,32 @@ import org.bukkit.command.CommandSender;
 import java.util.List;
 
 public interface ICommand {
-    //    public MinigamePlayerManager playerManager = Minigames.plugin.getPlayerData();
-//    public MinigameManager minigameManager = Minigames.plugin.getMinigameData();
     Minigames plugin = Minigames.getPlugin();
 
+    /**
+     * get the name of this (sub) command
+     * the name get used as it's first parameter
+     * in case of the command /mg the set subcommand /mg set implements this interface,
+     * as well as every sub-subcommand of it. set is the name of this specific subcommand
+     * and the name of all the ones one layer down are its parameters.
+     */
     String getName();
 
+    /**
+     * Get the aliases to the name of this command.
+     * An alias is an alternative name that should also show up as a parameter
+     * but calls the same command.
+     */
     String[] getAliases();
 
+    /**
+     * If the command can be called from the console or if it is strictly only callable by a player
+     */
     boolean canBeConsole();
 
+    /**
+     * Used in help command to describe what this (sub)command does.
+     */
     String getDescription();
 
     String[] getParameters();
