+++ Minigames/src/main/java/au/com/mineauz/minigames/recorder/BasicRecorder.java
@@ -1,4 +1,4 @@
-package au.com.mineauz.minigames.blockRecorder;
+package au.com.mineauz.minigames.recorder;
 
 import au.com.mineauz.minigames.Minigames;
 import au.com.mineauz.minigames.managers.MinigamePlayerManager;
@@ -35,13 +35,20 @@ import org.bukkit.inventory.InventoryHolder;
 
 /**
  * This class records all relevant changes the player does in the world while in a minigame
+ * After the minigame ends (or the player leaves it) this changes will be reverted.
+ * Shot arrows will be removed, blocks will be placed again, etc.
  */
 public class BasicRecorder implements Listener {
-
     private final MinigamePlayerManager playerManager = Minigames.getPlugin().getPlayerManager();
 
-    //helper methode to not write code double
-    //this checks for Minigame signs and cancels the event, adds the block to the block recorder and handles block drops
+    /**
+     * helper methode for blockBreak event to not write code double
+     * this checks for Minigame signs and cancels the event, adds the block to the block recorder and handles block drops
+     *
+     * @param event    the fired block break event
+     * @param mgPlayer the minigame representation of a player who broke the block
+     * @return true if the BlockBreakEvent should be canceled - aka if a minigame sign would get broken or if block drops are not allowed
+     */
     private boolean handleBlockBreak(BlockBreakEvent event, MinigamePlayer mgPlayer) {
         Block eBlock = event.getBlock();
         Minigame mgm = mgPlayer.getMinigame();
@@ -61,6 +68,7 @@ public class BasicRecorder implements Listener {
             }
         }
 
+        // signs are safe. Now check for block drops
         recData.addBlock(event.getBlock(), mgPlayer);
         if (!mgm.canBlocksdrop()) {
             event.getBlock().setType(Material.AIR);
@@ -70,7 +78,9 @@ public class BasicRecorder implements Listener {
         return false;
     }
 
-    //replace broken blocks (and stop breaking of blocks the player is not allowed to) <-- todo move this stopping into another more fitting class
+    /**
+     * record blocks broken by players (and stop breaking of blocks the player is not allowed to)
+     */
     @EventHandler(priority = EventPriority.HIGH, ignoreCancelled = true)
     private void blockBreak(BlockBreakEvent event) {
         MinigamePlayer mgPlayer = playerManager.getMinigamePlayer(event.getPlayer());
@@ -109,7 +119,9 @@ public class BasicRecorder implements Listener {
         }
     }
 
-    //remove placed blocks (and stop placing ones the player is not allowed to) <-- todo move this stopping into another more fitting class
+    /**
+     * record placed blocks (and stop placing ones the players are not allowed to)
+     */
     @EventHandler(priority = EventPriority.HIGH, ignoreCancelled = true)
     private void blockPlace(BlockPlaceEvent event) {
         MinigamePlayer mgPlayer = playerManager.getMinigamePlayer(event.getPlayer());
@@ -141,7 +153,9 @@ public class BasicRecorder implements Listener {
         }
     }
 
-    //reset changed inventories
+    /**
+     * record inventories changed by players
+     */
     @EventHandler(priority = EventPriority.HIGH, ignoreCancelled = true)
     private void changeAnotherInventory(PlayerInteractEvent event) {
         MinigamePlayer mgPlayer = playerManager.getMinigamePlayer(event.getPlayer());
@@ -158,32 +172,9 @@ public class BasicRecorder implements Listener {
         }
     }
 
-    //re-add drained liquid
-    @EventHandler(priority = EventPriority.HIGH, ignoreCancelled = true)
-    private void bucketFill(PlayerBucketFillEvent event) {
-        MinigamePlayer mgPlayer = playerManager.getMinigamePlayer(event.getPlayer());
-
-        if (mgPlayer.isInMinigame()) {
-            Minigame mgm = mgPlayer.getMinigame();
-            RecorderData recData = mgm.getRecorderData();
-
-            if (mgm.canBlockBreak()) {
-                if (recData.getWhitelistMode()) {
-                    //white list --> blocks that are allowed to be broken
-                    if (recData.getWBBlocks().contains(event.getBlock().getType())) {
-                        recData.addBlock(event.getBlockClicked(), playerManager.getMinigamePlayer(event.getPlayer()));
-                    }
-                    //black list --> blocks that are not allowed to be broken
-                } else if (!recData.getWBBlocks().contains(event.getBlock().getType())) {
-                    recData.addBlock(event.getBlockClicked(), playerManager.getMinigamePlayer(event.getPlayer()));
-                }
-            } else {
-                event.setCancelled(true);
-            }
-        }
-    }
-
-    //remove lighted fire
+    /**
+     * record fire lighted by players
+     */
     @EventHandler(priority = EventPriority.HIGH, ignoreCancelled = true)
     public void igniteBlock(BlockIgniteEvent event) {
         if (event.getPlayer() != null) {
@@ -211,7 +202,9 @@ public class BasicRecorder implements Listener {
         }
     }
 
-    //remove placed hanging entities
+    /**
+     * record hanging entities like item frames placed by players
+     */
     @EventHandler(priority = EventPriority.HIGH, ignoreCancelled = true)
     private void hangingPlace(HangingPlaceEvent event) {
         if (event.getPlayer() != null) {
@@ -233,7 +226,6 @@ public class BasicRecorder implements Listener {
                     usedMaterial = event.getItemStack().getType();
                 }
 
-                //todo we might want a flag parallel to BlockPlace/break for entities spawn / hurt
                 if (recData.getWhitelistMode()) {
                     //white list --> blocks that are allowed to be broken
                     if (recData.getWBBlocks().contains(usedMaterial)) {
@@ -251,10 +243,12 @@ public class BasicRecorder implements Listener {
         }
     }
 
-    //revive killed animals
-    //entities like monsters or armor stands are reset if restore regions is turned on.
+    /**
+     * record animals killed by players
+     * Note: entities like monsters or armor stands getting recorded only if regen regions are getting used
+     */
     @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
-    private void animalHurt(EntityDamageByEntityEvent event) { //todo check for spawn to not revive animals spawned while the minigame was running.
+    private void animalHurt(EntityDamageByEntityEvent event) {
         if (event.getEntity() instanceof Animals animal) {
             //did the player kill the animal?
             if (animal.getHealth() <= event.getDamage()) {
@@ -277,7 +271,9 @@ public class BasicRecorder implements Listener {
         }
     }
 
-    //revive hanging entities
+    /**
+     * record hanging entities getting removed by a player
+     */
     @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
     private void hangingBreak(HangingBreakByEntityEvent event) {
         MinigamePlayer mgPlayer = null;
@@ -296,7 +292,9 @@ public class BasicRecorder implements Listener {
         }
     }
 
-    //remove arrow entities
+    /**
+     * record arrows shot by players
+     */
     @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
     private void arrowShoot(EntityShootBowEvent event) {
         if (event.getEntity() instanceof Player player) {
@@ -308,7 +306,9 @@ public class BasicRecorder implements Listener {
         }
     }
 
-    //remove ender perl entities (important if some where caught in bubble columns)
+    /**
+     * record ender perl entities (important if somewhere caught in bubble columns)
+     */
     @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
     private void throwEnderPearl(ProjectileLaunchEvent event) {
         if (event.getEntity().getShooter() instanceof Player player) {
@@ -320,8 +320,36 @@ public class BasicRecorder implements Listener {
         }
     }
 
-    //remove liquid source
-    //todo this doesn't handle changed blocks of flowing liquids, as well as new created water sources
+    /**
+     * record liquid drained by players via bucket fill
+     */
+    @EventHandler(priority = EventPriority.HIGH, ignoreCancelled = true)
+    private void bucketFill(PlayerBucketFillEvent event) {
+        MinigamePlayer mgPlayer = playerManager.getMinigamePlayer(event.getPlayer());
+
+        if (mgPlayer.isInMinigame()) {
+            Minigame mgm = mgPlayer.getMinigame();
+            RecorderData recData = mgm.getRecorderData();
+
+            if (mgm.canBlockBreak()) {
+                if (recData.getWhitelistMode()) {
+                    //white list --> blocks that are allowed to be broken
+                    if (recData.getWBBlocks().contains(event.getBlock().getType())) {
+                        recData.addBlock(event.getBlockClicked(), playerManager.getMinigamePlayer(event.getPlayer()));
+                    }
+                    //black list --> blocks that are not allowed to be broken
+                } else if (!recData.getWBBlocks().contains(event.getBlock().getType())) {
+                    recData.addBlock(event.getBlockClicked(), playerManager.getMinigamePlayer(event.getPlayer()));
+                }
+            } else {
+                event.setCancelled(true);
+            }
+        }
+    }
+
+    /**
+     * record liquid placed by players via bukkit empty
+     */
     @EventHandler(priority = EventPriority.HIGH, ignoreCancelled = true)
     private void bucketEmpty(PlayerBucketEmptyEvent event) {
         MinigamePlayer mgPlayer = playerManager.getMinigamePlayer(event.getPlayer());
@@ -354,7 +382,9 @@ public class BasicRecorder implements Listener {
         }
     }
 
-    //revive vehicles
+    /**
+     * record vehicles broken by players
+     */
     @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
     private void vehicleDestroy(VehicleDestroyEvent event) {
         if (event.getAttacker() != null) {
