+++ Minigames/src/main/java/au/com/mineauz/minigames/recorder/RecorderData.java
@@ -1,4 +1,4 @@
-package au.com.mineauz.minigames.blockRecorder;
+package au.com.mineauz.minigames.recorder;
 
 import au.com.mineauz.minigames.Minigames;
 import au.com.mineauz.minigames.menu.Callback;
@@ -31,6 +31,9 @@ import java.io.*;
 import java.lang.reflect.Type;
 import java.util.*;
 
+/**
+ *
+ */
 public class RecorderData implements Listener {
     // list of blocks that need another block to not break
     private static final ArrayList<Material> supportedMats = new ArrayList<>();
