+++ Minigames/src/main/java/au/com/mineauz/minigames/recorder/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * recorder is the system for recording block / entity changes in a game and rolling them back afterward.
+ */
+package au.com.mineauz.minigames.recorder;
