+++ Minigames/src/main/resources/plugin.yml
@@ -2,7 +2,7 @@ name: Minigames
 main: au.com.mineauz.minigames.Minigames
 version: ${plugin.version}
 api-version: ${spigot.api.version}
-author: _Razz_ Schmoller Addstar Narimm
+author: _Razz_ Schmoller Addstar Narimm GreenSurvivors
 softdepend: [ 'Multiverse-Core' ]
 description: ${project.description}
